+++ src/main/java/mod/akrivus/amalgam/client/render/layers/LayerNephriteItem.java
@@ -1,8 +1,6 @@
 package mod.akrivus.amalgam.client.render.layers;
 
 import mod.akrivus.kagic.client.render.RenderGemBase;
-import mod.akrivus.kagic.client.render.RenderHoloPearl;
-import mod.akrivus.kagic.client.render.RenderPearl;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.ModelBiped;
 import net.minecraft.client.renderer.GlStateManager;
@@ -45,7 +43,7 @@ public class LayerNephriteItem implements LayerRenderer<EntityLivingBase> {
 			GlStateManager.rotate(-90.0F, 1.0F, 0.0F, 0.0F);
 			GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
 			boolean flag = handSide == EnumHandSide.LEFT;
-			GlStateManager.translate((float)(flag ? -1 : 1) / 16.0F, 0.125F, -0.7F);
+			GlStateManager.translate((flag ? -1 : 1) / 16.0F, 0.125F, -0.7F);
 			Minecraft.getMinecraft().getItemRenderer().renderItemSide(entity, stack, camera, flag);
 			GlStateManager.popMatrix();
 		}
