+++ src/main/java/mod/akrivus/amalgam/client/render/layers/LayerStevonnieItem.java
@@ -1,6 +1,5 @@
 package mod.akrivus.amalgam.client.render.layers;
 
-import mod.akrivus.amalgam.client.render.RenderConnie;
 import mod.akrivus.amalgam.client.render.RenderStevonnie;
 import mod.akrivus.amalgam.gem.EntityStevonnie;
 import net.minecraft.client.Minecraft;
@@ -43,7 +42,7 @@ public class LayerStevonnieItem implements LayerRenderer<EntityStevonnie> {
 			GlStateManager.rotate(-90.0F, 1.0F, 0.0F, 0.0F);
 			GlStateManager.rotate(180.0F, 0.0F, 1.0F, 0.0F);
 			boolean flag = handSide == EnumHandSide.LEFT;
-			GlStateManager.translate((float)(flag ? -1 : 1) / 6.0F, 0.125F, -0.75F);
+			GlStateManager.translate((flag ? -1 : 1) / 6.0F, 0.125F, -0.75F);
 			Minecraft.getMinecraft().getItemRenderer().renderItemSide(entity, stack, camera, flag);
 			GlStateManager.popMatrix();
 		}
